Define relationship between every model

+++ db/migrate/20210624072952_create_menu_items.rb
@@ -1,10 +1,10 @@
 class CreateMenuItems < ActiveRecord::Migration[6.1]
   def change
     create_table :menu_items do |t|
-      t.integer :menu_categorie_id
+      t.bigint :menu_categorie_id
       t.string :name
       t.string :description
-      t.integer :price
+      t.float :price
       t.boolean :status
     end
   end
